requirements.txt
openai>=1.0.0
anthropic>=0.20.0
pyyaml>=6.0
praw>=7.7.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
